Isolating and organizing the nag dialogs
