+++ MekHQ/src/mekhq/gui/CampaignGUI.java
@@ -94,9 +94,9 @@ import mekhq.gui.dialog.DataLoadingDialog;
 import mekhq.gui.dialog.GMToolsDialog;
 import mekhq.gui.dialog.HireBulkPersonnelDialog;
 import mekhq.gui.dialog.HistoricalDailyReportDialog;
-import mekhq.gui.dialog.InsufficientAstechTimeNagDialog;
-import mekhq.gui.dialog.InsufficientAstechsNagDialog;
-import mekhq.gui.dialog.InsufficientMedicsNagDialog;
+import mekhq.gui.dialog.nagDialogs.InsufficientAstechTimeNagDialog;
+import mekhq.gui.dialog.nagDialogs.InsufficientAstechsNagDialog;
+import mekhq.gui.dialog.nagDialogs.InsufficientMedicsNagDialog;
 import mekhq.gui.dialog.MaintenanceReportDialog;
 import mekhq.gui.dialog.MassMothballDialog;
 import mekhq.gui.dialog.MekHQAboutBox;
@@ -105,7 +105,7 @@ import mekhq.gui.dialog.MekHqOptionsDialog;
 import mekhq.gui.dialog.MercRosterDialog;
 import mekhq.gui.dialog.NewRecruitDialog;
 import mekhq.gui.dialog.NewsReportDialog;
-import mekhq.gui.dialog.OutstandingScenariosNagDialog;
+import mekhq.gui.dialog.nagDialogs.OutstandingScenariosNagDialog;
 import mekhq.gui.dialog.PartsStoreDialog;
 import mekhq.gui.dialog.PersonnelMarketDialog;
 import mekhq.gui.dialog.PopupValueChoiceDialog;
@@ -114,11 +114,11 @@ import mekhq.gui.dialog.ReportDialog;
 import mekhq.gui.dialog.RetirementDefectionDialog;
 import mekhq.gui.dialog.ScenarioTemplateEditorDialog;
 import mekhq.gui.dialog.ShipSearchDialog;
-import mekhq.gui.dialog.ShortDeploymentNagDialog;
+import mekhq.gui.dialog.nagDialogs.ShortDeploymentNagDialog;
 import mekhq.gui.dialog.UnitCostReportDialog;
 import mekhq.gui.dialog.UnitMarketDialog;
-import mekhq.gui.dialog.UnmaintainedUnitsNagDialog;
-import mekhq.gui.dialog.UnresolvedStratConContactsNagDialog;
+import mekhq.gui.dialog.nagDialogs.UnmaintainedUnitsNagDialog;
+import mekhq.gui.dialog.nagDialogs.UnresolvedStratConContactsNagDialog;
 import mekhq.gui.model.PartsTableModel;
 import mekhq.io.FileType;
 import org.w3c.dom.Document;
