+++ MekHQ/src/mekhq/gui/dialog/nagDialogs/InsufficientMedicsNagDialog.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with MekHQ. If not, see <http://www.gnu.org/licenses/>.
  */
-package mekhq.gui.dialog;
+package mekhq.gui.dialog.nagDialogs;
 
 import mekhq.MekHQ;
 import mekhq.MekHqConstants;
